requirements.txt
Scrapy
scrapy-user-agents
discord.py
openpyxl
tabulate
xlsxwriter
python-dateutil